Use ASSETS_DEBUG rather than DEBUG for using staticfiles
Django 1.6.x makes the default test-runner set DEBUG=False
during test initialization. As a result, the Django Assets
cannot be found during the test-run when rendering templates.

This change simply switches to using ASSETS_DEBUG to control
the behaviour, with DEBUG used only if ASSETS_DEBUG is not
defined in settings.